07_authorization_in_flask_and_react/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
appnope==0.1.4
asttokens==2.4.1
backcall==0.2.0
bcrypt==4.2.0
blinker==1.8.2
cachelib==0.13.0
click==8.1.7
decorator==5.1.1
executing==2.0.1
Faker==27.0.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
ipdb==0.13.13
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
msgspec==0.18.6
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==3.0.3
zipp==3.20.0
